backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-generativeai>=0.8.0
python-dotenv==1.0.0
httpx==0.25.2
geopy==2.4.0
PyJWT==2.8.0
psycopg2-binary==2.9.9
